distributed dashboard
