+++ dashboard/v1.0/src/components/dashboard/MonitoringSummary.tsx
@@ -1,0 +1,5 @@
+import React, { FC } from 'react';
+
+export const MonitoringSummary: FC<{}> = () => {
+  return <div className="col-md-6"></div>;
+};
